requirements.txt
fastapi
uvicorn
python-dotenv
google-generativeai
langchain-community
langchain-core
chromadb
sentence-transformers
langchain
requests
datasets
torch==2.9.1
transformers==4.57.2
sentencepiece
-f https://download.pytorch.org/whl/cpu/torch_stable.html
